Split job submit from job state
The nersc / jaws flow manager has a dependency on job state as it'll
need to get and update the state. The job submission code will have a
dependency on the ners / jaws flow manager, and so needs to be split out
to avoid a circular import.
