+++ cdmtaskservice/app_state.py
@@ -18,6 +18,7 @@ from cdmtaskservice.image_remote_lookup import DockerImageInfo
 from cdmtaskservice.images import Images
 from cdmtaskservice.jobflows.nersc_jaws import NERSCJAWSRunner
 from cdmtaskservice.job_state import JobState
+from cdmtaskservice.job_submit import JobSubmit
 from cdmtaskservice.kb_auth import KBaseAuth
 from cdmtaskservice.mongo import MongoDAO
 from cdmtaskservice.nersc.client import NERSCSFAPIClientProvider
@@ -34,6 +35,7 @@ class AppState(NamedTuple):
     auth: KBaseAuth
     sfapi_client: NERSCSFAPIClientProvider
     s3_client: S3Client
+    job_submit: JobSubmit
     job_state: JobState
     images: Images
     # TODO CODE make an abstract jobflow class or something. For now just duck type them
@@ -86,7 +88,7 @@ async def build_app(
     logr.info("Done")
     try:
         mongodao = await MongoDAO.create(mongocli[cfg.mongo_db])
-        job_state = JobState(mongodao, s3)
+        job_state = JobState(mongodao)
         nerscjawsflow = NERSCJAWSRunner(
             nerscman,
             job_state,
@@ -100,9 +102,10 @@ async def build_app(
         runners = {models.Cluster.PERLMUTTER_JAWS: nerscjawsflow}
         imginfo = await DockerImageInfo.create(Path(cfg.crane_path).expanduser().absolute())
         images = Images(mongodao, imginfo)
+        job_submit = JobSubmit(mongodao, s3)
         app.state._mongo = mongocli
         app.state._cdmstate = AppState(
-            auth, sfapi_client, s3, job_state, images, runners
+            auth, sfapi_client, s3, job_submit, job_state, images, runners
         )
     except:
         mongocli.close()
