+++ cdmtaskservice/error_mapping.py
@@ -9,7 +9,7 @@ from cdmtaskservice.errors import ErrorType
 from cdmtaskservice.http_bearer import MissingTokenError, InvalidAuthHeaderError
 from cdmtaskservice.images import NoEntrypointError
 from cdmtaskservice.image_remote_lookup import ImageNameParseError, ImageInfoFetchError
-from cdmtaskservice.job_state import ETagMismatchError
+from cdmtaskservice.job_submit import ETagMismatchError
 from cdmtaskservice.kb_auth import InvalidTokenError, MissingRoleError
 from cdmtaskservice.mongo import (
     ImageTagExistsError,
