+++ test/job_submit_test.py
@@ -1,0 +1,7 @@
+# TODO TEST add tests
+
+from cdmtaskservice import job_submit  # @UnusedImport
+
+
+def test_noop():
+    pass
